Split up wrapitup into multiple files for readability
