+++ tests/__init__.py
@@ -1,0 +1,8 @@
+"""Tests for WrapItUp.
+
+Execute the tests with::
+
+	$ python -m unittest discover tests
+
+from the root directory of the source repository.
+"""
